Select specific alloy(s) and display specs data
Added the capability to select one OR two specific alloy(s) and display the corresponding specs data.
The user selects between the two options using an appropriate checkbox.
In the case where two alloys are selected, the difference between the specs of the next and the previous alloy is also displayed along with their individual specs.

Removed the option to display all alloy data in a single table because it doesn't seem to be of use, at least for the time being.