Add "Production Schedule" app
Added part of a whole new "Production Schedule" app. The new app is accessible by a sidebar selectbox alongside the existing "Change Alloy" app.

Added the first function of the new app, also accessible by a sidebar selectbox. This function displays orders in a tabular form, grouped by the Casthouse section at which they are scheduled to be cast.

Added everything needed to connect to a Firestore database, used to store the order data (for now). The related key is locally stored in a file included in  .gitignore. For deployment purposes, the same information is stored in st.secrets via Streamlit Sharing.