Add "Product info" function
Added a "Product info" function to "Production Schedule" app.

Information about the various products is available either by an expandable table containing all products or in the form of a table containing a single product, chosen by the user via a selectbox.

Of course, this was preceded by adding the capability to read product data stored in Dropbox as a csv file, using the Dropbox API.