Improve date handling
Added an if-clause to prevent the user from adding or editing an order in a way that causes the due date to be earlier than the start date.

Added a visual cue to inform the user that the due date of a specific order has expired at the production schedule  table display.
More specifically, the corresponding due date is written in red letters.

Minor: Made it so that "Production Schedule" app is now the default instead of "Change Alloy" app.

Fix: Fixed a bug related to start and due dates in "Edit order" function.